fix SpriteView in simulator
When a css background-image is set that has an etag and no max-age (as when developing using the browser simulator in debug mode), Chrome sends an HTTP request for images to check for a 304 response, causing the background image to flicker.  When using a sprite-view updating every frame, this causes tons of http requests and flickering.  To get around this, we add a separate DOM node for each unique sprite sheet that an image can show (opts['dom:multipleImageNodes'] == true).  The SpriteView (subclass of ImageView) enables this option by default when we're in debug mode.