Fix invalid composite op warnings from PEngine
The native view implementation assumes that when compositeOperation has
a string value, even an empty one, that it should try resolve the value
to some compositeOperation. Empty strings don't map to anything there.
Passing null in allows a bunch of branching logic to be skipped and
avoids log spam from clearing a composite operation.