change how input event capturing works
Previously, timestep captured all input regardless of whether it
originated on the timestep canvas.  Now, if you place dom nodes over the
canvas, you can interact with them and the touch events will not
propagate to teh game underneath.  This allows building DOM UI for
browser-based games.  If you are dragging a finger that started on the
canvas, however, the canvas will continue to receive events regardless
of where your finger is until the finger is released.