cleanup (imagescaleview): remove urlChanged listeners
Remove urlChanged listeners in imagescaleview - nothing in timestep
emits urlChanged, so this was a bit of a hack to allow monkeypatching
refreshes from outside the library by emitting events from the underlying
image objects. This is not needed to make the imagescaleview work correctly,
so it is being removed. If an external way to trigger a refresh is needed
in the future, imagescaleview should expose that as an api.