ImageScaleView: reduce slices larger than the image
In the specific case where the middle slice is inferred from the
surrounding slices and the surrounding slices are larger than the size
of the image, the middle slice ends up negative, causing a rendering
glitch.  Instead, adjust the edge slices to allow for a middle slice
with a minimum size of 1.