enable scrolling on landscape ios >= 7
allows scrolling out the address bar when in an iphone mobile browser in
the landscape orientation

By default we allow landscape scrolling in iOS 7+, but iOS 7 does not
have friendly UX for landscape since tapping anywhere near the top or
bottom of the screen will use make the navbar appear.  Thus, for many
games, it is desirable to leave the navbar open all the time on iOS 7
(even though the viewport will be very small), disabling scrolling
completely.  To disable scrolling, use
`Application._settings.minIOSLandscapeScroll` to set the min version for
which landscape scrolling.  Set `disableIOSLandscapeScroll` to disable
this feature completely.

src/Application:

  exports = Class(ui.View, function () {
    // disable ios landscape scrolling in iOS 7, only allow 8+
    this._settings = {
      minIOSLandscapeScroll: 8
    };
  });

Note that landscape scrolling is only relevant for web builds (iOS
mobile safari) on iPhones (or iPods), not iPads and not standalone web
browsers (e.g. home screen website or embedded web view).